地図コンポーネントを分離リファクタ
